fix: memory safety and bounds validation fixes
- console.cpp: check format[i] for '\0' after i++ on lone '%'
- include/core/memory.h: switch kmalloc/kbrk/krealloc from int to size_t
- core/memory.cpp: switch kmalloc/kbrk/krealloc/worst_fit/allocate_new_block
  from signed int to size_t; drop (int) cast in kbrk comparison; add
  INT_MAX overflow guard in kcalloc before truncating to int
- core/driver.cpp: update EXPORT_SYMBOL_ASM mangled name for kmalloc(size_t)
- core/drivers/keyboard.cpp: remove WaitForKBACK after controller commands
  (0xAE/0x60); poll output-buffer-ready with timeout before read; fix
  Shift+Caps — XOR for letters, apply shift for non-letters
- core/drivers/ModuleLoader.cpp: reject internal symbols whose section
  lacks SHF_ALLOC or whose value exceeds section size; same SHF_ALLOC+size
  validation for CreateDriverInstance entry-point
- core/filesystem/FAT32.cpp: validate BPB geometry (sectorsPerCluster pow2,
  fatCopies>0, tableSize>0, rootCluster in [2,maxCluster)) before mounting;
  add fat_valid_cluster() helper and replace bare `==0` chain checks with
  range-guarded validation across all FAT walkers
- core/syscalls.cpp: replace lidt-from-address-0 with packed null IDTR
  struct for deterministic triple-fault
- gui/button.cpp: add multiplication overflow guard in SetWidth/SetHeight
- gui/terminalview.cpp: add null/zero-dimension guard in RedrawToCache
- gui/widget.cpp: deliver OnMouseMove to captured/pressed children
- Makefile: add .PHONY declarations for build/check targets

+++ include/core/memory.h
@@ -40,7 +40,7 @@ int kheap_init(void* start_addr, void* end_addr);
 /**
  * increase the heap memory by size & get its address
  */
-void* kbrk(int size);
+void* kbrk(size_t size);
 
 /**
  * print list of allocated blocks
@@ -52,7 +52,7 @@ void kheap_print_blocks();
  * otherwise try some memory allocation algorithm like best fit etc
  * to find best block to allocate
  */
-void* kmalloc(int size);
+void* kmalloc(size_t size);
 
 void* aligned_kmalloc(size_t size, size_t alignment);
 void aligned_kfree(void* ptr);
@@ -66,7 +66,7 @@ void* kcalloc(int n, int size);
  * allocate a new block of memory
  * copy previous block data & set free the previous block
  */
-void* krealloc(void* ptr, int size);
+void* krealloc(void* ptr, size_t size);
 
 /**
  * set free the block
